refactor: separate lexer and parser tests
